feat: introduce system router for health and platforms endpoints, enhancing API organization and adding tests
