+++ backend-migration/app/framework/api/endpoint_registry.py
@@ -4,6 +4,7 @@ from fastapi import FastAPI
 
 from app.api.endpoints import metadata
 from app.framework.api.debug_router import create_debug_router
+from app.framework.api.system_router import create_system_router
 
 
 def register_default_endpoints(app: FastAPI) -> None:
@@ -15,3 +16,4 @@ def register_default_endpoints(app: FastAPI) -> None:
     """
     app.include_router(metadata.router)
     app.include_router(create_debug_router())
+    app.include_router(create_system_router())
